make the queueSize percept also work for carriers and reavers
