+++ eisbw/src/main/java/eisbw/units/StarcraftUnitFactory.java
@@ -3,7 +3,6 @@ package eisbw.units;
 import java.util.ArrayList;
 import java.util.List;
 
-import eisbw.percepts.perceivers.BuildingPerceiver;
 import eisbw.percepts.perceivers.GenericUnitPerceiver;
 import eisbw.percepts.perceivers.IPerceiver;
 import jnibwapi.JNIBWAPI;
@@ -34,11 +33,8 @@ public class StarcraftUnitFactory {
 	 * @return - a StarCraft unit with perceivers.
 	 */
 	public StarcraftUnit create(Unit unit) {
-		List<IPerceiver> perceptGenerators = new ArrayList<>(2);
+		List<IPerceiver> perceptGenerators = new ArrayList<>(1);
 		perceptGenerators.add(new GenericUnitPerceiver(this.api, unit));
-		if (unit.getType().isBuilding()) {
-			perceptGenerators.add(new BuildingPerceiver(this.api, unit));
-		}
 		return new StarcraftUnit(perceptGenerators, unit.getType().isWorker());
 	}
 }
